Expose metadata via `DataAPI.metadata`
Use the `metadata` function recently added to DataAPI to expose
the contents of `ReadStatMeta`. This will allow e.g. `DataFrame` to import
the metadata.

The implementation is not super efficient as a new `Dict` is allocated for
each column. If this turns out to be a problem it would be possible to
create a special `AbstractDict` type that would expose the data without
allocating, but this may be overkill. Also, if users mutate the returned
metadata `Dict`s, the changes will have no effect on `ReadStatMeta`,
except if they mutate the `Dicts` holding value labels.